requirements.txt
asgiref==3.10.0
Django==5.2.8
django-graphql-jwt==0.4.0
djangorestframework==3.16.1
graphene==3.4.3
graphene-django==3.2.3
graphene-file-upload==1.3.0
graphql-core==3.2.7
graphql-relay==3.2.0
pillow==12.0.0
promise==2.3
PyJWT==2.10.1
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.15.0
